OLMIS-5986: For shipment files that did not process successfully, write an error log file in the errors folder that describes why it failed to process.

+++ src/main/java/org/openlmis/fulfillment/service/shipment/ShipmentMessageHandler.java
@@ -16,7 +16,9 @@
 package org.openlmis.fulfillment.service.shipment;
 
 import java.io.File;
+import java.io.FileNotFoundException;
 import java.io.IOException;
+import java.io.PrintWriter;
 import java.nio.file.Files;
 import java.util.List;
 import javax.transaction.Transactional;
@@ -26,6 +28,7 @@ import org.openlmis.fulfillment.domain.FileTemplate;
 import org.openlmis.fulfillment.domain.Shipment;
 import org.openlmis.fulfillment.domain.TemplateType;
 import org.openlmis.fulfillment.service.FileTemplateService;
+import org.openlmis.fulfillment.service.FulfillmentException;
 import org.openlmis.fulfillment.service.ShipmentService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -57,6 +60,9 @@ public class ShipmentMessageHandler {
   @Autowired
   private ApplicationContext context;
 
+  @Autowired
+  private ShipmentArchiveFileNameGenerator fileNameGenerator;
+
   /**
    * A message handler endpoint that processes incoming shipment files.
    *
@@ -74,11 +80,46 @@ public class ShipmentMessageHandler {
       shipmentService.save(shipment);
       archiveFile(message, "outboundShipmentFileArchiveChannel");
     } catch (RuntimeException exception) {
-      LOGGER.warn(exception.getLocalizedMessage(), exception);
-      archiveFile(message, "errorChannel");
+      logException(file, exception);
     } finally {
-      Files.delete(file.toPath());
+      if (file.exists()) {
+        Files.delete(file.toPath());
+      }
     }
+  }
+
+  private void logException(File file, RuntimeException exception) throws IOException {
+    String prefix = fileNameGenerator.generatePrefix();
+    String fileName = prefix + file.getName();
+
+    File originalPayload = new File(
+        file.getParentFile().getParentFile().getPath().concat(File.separator).concat(fileName));
+    Files.move(file.toPath(), originalPayload.toPath());
+
+    Message<File> payload = MessageBuilder.withPayload(originalPayload).build();
+    archiveFile(payload, "errorChannel");
+    Files.delete(originalPayload.toPath());
+
+    File errorLogFile = writeExceptionToFile(exception, file
+        .getParentFile().getPath().concat(File.separator).concat(fileName));
+    Message<File> errorLog = MessageBuilder.withPayload(errorLogFile).build();
+
+    archiveFile(errorLog, "errorChannel");
+    Files.delete(errorLogFile.toPath());
+  }
+
+  private File writeExceptionToFile(RuntimeException exception, String fileName)
+      throws FileNotFoundException {
+    File errorLogFile = new File(fileName + ".log");
+    PrintWriter writer = new PrintWriter(errorLogFile);
+    if (exception instanceof FulfillmentException) {
+      writer.println(exception.getMessage());
+    } else {
+      exception.printStackTrace(writer);
+    }
+    writer.flush();
+    writer.close();
+    return errorLogFile;
   }
 
   private void archiveFile(Message<File> message, String archiveFtpChannel) {
